XITE_GNN/requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
gatspy==0.3
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.2
nvidia-ml-py==11.525.84
nvitop==1.0.0
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.4
py-ecg-detectors==1.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
termcolor==2.2.0
threadpoolctl==3.1.0
torch==1.10.0+cu111
torch-geometric==2.0.4
torch-scatter==2.1.0
torch-sparse==0.6.16
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
